Add SWO Command Handler Implementations
This adds command handlers for the DAP SWO functions, but does not yet
mark this probe firmware as supporting SWO, since it is currently
difficult to test out this functionality with the state of currently
released host debugging software.

Eventually this will need to be re-visited and the capabilities needs
to indicate that SWO is supported, once it is tested.